+++ CHANGELOG.md
@@ -3,5 +3,6 @@
 * Creation of the Environment Detector service.
 * Creation of the ErrorWrapper abstraction and the InvalidParameterError.
 * Creation of the the ConfigFileManager and the ConfigFileError.
-* Creation of the AppConfigurator, ConfigData abstraction and the ConfiguratorError.
+* Creation of the AppConfigurator and ConfigData abstraction.
 * Creation of the MainWindowController, MainWindowAlreadyInitializedError and MainWindowNotInitializedError.
+* Creation of the i18n engine, init object and i18n error object.